TransactionsServices-main/requirements.txt
kafka-python
customtkinter
requests